src/requirements.txt
# dev
black==24.10.0

# env
python-dotenv==1.0.1

# generation
tqdm==4.67.1

# evaluation
deepeval==2.0.5



